Move OidcConfigurationProvider to .oidc package
Issue: gh-5355
